requirements.txt
pjsua2
openai
elevenlabs
sounddevice
numpy
ffmpeg-python
pyyaml